Written a new test class TestGraphRollback with a large collection of tests, many of them are taken from existing ones in TestObjectContext.m. The new test class makes the testing code a lot more compact and easier to extend.
Improved the navigation of the context history to recreate temporal instances 
that exist at the requested version, but don't at the current one. This is one 
of the last important missing bit to properly support temporal navigation.
Also tried to limit the number of faults in the reverted object graph, 
by traversing the graph and resolving all UUIDs for which cached objects exist 
in memory. This is still under evaluation, because this feature is theorically 
not needed.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@3691 7b3f36aa-e6db-49c1-8b4c-66eac1997e64