Various small improvements and new utility methods in COObjectContext.
Improved context reverting to discard future objects that haven't yet been 
created at the wanted context version. Modified to scan the context history back 
to the initial version, this allows to revert the objects that aren't loaded but 
have to because they belong to the context. Until now, only loaded/registered 
objects were all properly reverted. Also modifed -objectForUUID:version: to call
 -commitMergeOfInstance:forObject:, this fixes the fact that the revert of 
objects not loaded wasn't recorded in their object bundle.

Corrected the loading of objects in -recreateObjectGraph of TestGraphRollback. 
Added tests to check the state of the object cache.

git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@3696 7b3f36aa-e6db-49c1-8b4c-66eac1997e64