Hacked a workaround for merging bugs that might arise when restoring the context with a custom merge policy, and also to limit harmless but spurious warnings that can be logged when a group is restored. Documented in details the issue and outlines what a proper fix could be. Reworked those warnings logged by -resolveFaults to be more explicit.
Increased the safety of context restorations by swapping restored object and 
current object in the cache before updating the object graph. This can help 
in case fault resolution is wrongly triggered on a border case: it ensures 
UUID references to the object will be resolved to the restored object as 
expected rather than the current object.

git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@3897 7b3f36aa-e6db-49c1-8b4c-66eac1997e64