Fixed -currentNodeDidChange: to check the store correctly (multiple stores can now be used at the same time).
Previously the code assumed a single store was in use at a time. For the test 
suite, multiple stores (using distinct UUIDs) are created sequentially, and 
distributed notifications can be received with some delay due to the run loop...So in some cases, a COCustomTrack can receive a notification posted by a 
previous store (and not from the current one) resulting in an assertion until 
now.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@7823 7b3f36aa-e6db-49c1-8b4c-66eac1997e64