Imported SQLClient and Performance libraries based on GNUstep SVN trunk r35732
For GNUstep, we will just introduce new dependencies on the GNUstep libraries, 
but on Mac OS X to make the compilation easier we use the versions imported into 
the CoreObject subdirectories. 

The SQLClient code imported is limited to the SQLite and PostgreSQL backends.
For Performance, the library is reduced to two classes, GSCache and GSTicker.


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@7840 7b3f36aa-e6db-49c1-8b4c-66eac1997e64