Fixed reference to UnitKit.xcodeproj (to match the new location)
Note: All tests pass in EtoileFoundation, CoreObject and EtoileUI.

Warning: If 'Run' succeeds, but produces not ouput when running a test suite 
scheme...
Due to the UnitKit.xcodeproj location change, 'ukrun' could need to reset as a 
tool in the Run section (for each test suite scheme e.g. TestCoreObject).


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/CoreObject@10376 7b3f36aa-e6db-49c1-8b4c-66eac1997e64