Turned off designated initializer warnings in COPrimitiveCollection
What matters is that we always call a superclass designated initializer and we return a valid collection, and that's what we do.

These warnings arise with iOS 8 / OS X 10.10 SDKs due NS_DESIGNATED_INITIALIZER macro in the Foundation headers.

Note: For collection classes, -initWithCoder: and -initWithObjects:count:  are not the only designated initializers marked with NS_DESIGNATED_INITIALIZER.
Looks like a dubious choice, especially when you consider it seems to go against the designated initializer rule, and make it easier to get initialization wrong in subclasses (ensuring multiple initialization code paths are symmetric is more complex than just using a single one).