Fix assertion failure in Benchmark when this branch (history-navigation-optimization) is merged with master.
-[COUndoTrack commitStoreTransaction:] no longer forces loading persistent roots in order to get lastTransactionID.

The problem -testGoToFirstCommitNodeToReloadManyDeletedPersistentRoots was hitting was, a context with
no loaded persistent roots was attempting to undo the deletion of many persistent roots. It got to the point
of committing the transaction, but had never loaded any of the persistent roots, because COUndoTrack
can commit a persistent root undeletion without loading the persistent root first.

What I did is modify COSQLiteStore to persistent root modifications when no transaction ID is provided,
to accomodate this scenario. If one is provided, it still has to be correct.